Refactor notion_icon class and related functions; enhance documentation, improve error handling, and add unit tests for Create and ConvertFromObject methods. function New-NotionPage working (icon,cover)! function Remove-NullValuesFromObject repaired

+++ source/Private/Remove-NullValuesFromObject.ps1
@@ -29,8 +29,14 @@ function Remove-NullValuesFromObject
             if ($null -ne $property.Value)
             {
                 Write-Debug "Processing property: $($property.Name): $($property.Value) `<$($property.value.GetType().Name)]"
+
                 if (($property.value -is [int]) -or ($property.value -is [int64]) -or ($property.value -is [double]) -or ($property.value -is [float]) -or ($property.value -is [string]) -or ($property.value -is [bool]) -or ($property.value -is [DateTime]) -or ($property.value -is [enum]))
                 {
+                    if (($property.Value -is [string]) -and ([string]::IsNullOrEmpty($property.Value)))
+                    {
+                        # Skip empty strings
+                        continue :loop
+                    }
                     $outputObject | Add-Member -MemberType NoteProperty -Name $property.Name -Value $property.Value
                 }
 
@@ -42,19 +48,31 @@ function Remove-NullValuesFromObject
                         continue :loop
                     }
                     $cleanedArray = @()
-                    foreach ($item in $property.Value)
+                    :cleanArrayLoop foreach ($item in $property.Value)
                     {
-                        if($item -is [array]){
-                            $cleanedArray += ,@($(Remove-NullValuesFromObject -InputObject $item))
+                        if ( $null -eq $item )
+                        {
+                            continue :cleanArrayLoop
                         }
-                        else{
+                        if ($item -is [array])
+                        {
+                            $cleanedArray += , @($(Remove-NullValuesFromObject -InputObject $item))
+                        }
+                        else
+                        {
                             $cleanedArray += $(Remove-NullValuesFromObject -InputObject $item)
                         }
+                    }
+                    if ( $cleanedArray.Count -eq 0 )
+                    {
+                        # Skip empty arrays
+                        continue :loop
                     }
                     $outputObject | Add-Member -MemberType NoteProperty -Name $property.Name -Value $cleanedArray
                 }
                 elseif ($property.Value -is [Object] ) 
                 {
+                    $nestedObject = $null
                     try
                     {
                         $nestedObject = Remove-NullValuesFromObject -InputObject $property.Value
